Allocation Profiler: Types for all allocations
Before this PR, we were missing the types for allocations in two cases:
1. allocations from codegen
2. allocations in gc_managed_realloc_

The second one is easy: those are always used for `buffer`s, right?

For the first one: we pass the type through to the allocation function,
so that we can log the allocation with the correct type.